+++ common/src/main/java/com/human/common/registry/init/item/HumanPlasticBlockItems.java
@@ -71,6 +71,48 @@ public class HumanPlasticBlockItems {
                 )
         );
 
+    public static final Map<DyeColor, Supplier<BlockItem>> DYE_COLOR_TO_PITTED_PLASTIC =
+        Collections.unmodifiableMap(
+            Arrays.stream(DyeColor.values())
+                .collect(
+                    Collectors.toMap(
+                        Function.identity(),
+                        dyeColor -> AVPBlockItems.register(
+                            dyeColor.getName() + "_pitted_plastic",
+                            HumanPlasticBlocks.DYE_COLOR_TO_PITTED_PLASTIC.get(dyeColor)
+                        )
+                    )
+                )
+        );
+
+    public static final Map<DyeColor, Supplier<BlockItem>> DYE_COLOR_TO_PITTED_PLASTIC_SLAB =
+            Collections.unmodifiableMap(
+                    Arrays.stream(DyeColor.values())
+                            .collect(
+                                    Collectors.toMap(
+                                            Function.identity(),
+                                            dyeColor -> AVPBlockItems.register(
+                                                    dyeColor.getName() + "_pitted_plastic_slab",
+                                                    HumanPlasticBlocks.DYE_COLOR_TO_PITTED_PLASTIC_SLAB.get(dyeColor)
+                                            )
+                                    )
+                            )
+            );
+
+    public static final Map<DyeColor, Supplier<BlockItem>> DYE_COLOR_TO_PITTED_PLASTIC_STAIRS =
+            Collections.unmodifiableMap(
+                    Arrays.stream(DyeColor.values())
+                            .collect(
+                                    Collectors.toMap(
+                                            Function.identity(),
+                                            dyeColor -> AVPBlockItems.register(
+                                                    dyeColor.getName() + "_pitted_plastic_stairs",
+                                                    HumanPlasticBlocks.DYE_COLOR_TO_PITTED_PLASTIC_STAIRS.get(dyeColor)
+                                            )
+                                    )
+                            )
+            );
+
     public static final Map<DyeColor, Supplier<BlockItem>> DYE_COLOR_TO_PLASTIC =
         Collections.unmodifiableMap(
             Arrays.stream(DyeColor.values())