+++ fabric/src/main/java/com/avp/fabric/data/tag/AVPBlockTagProvider.java
@@ -176,6 +176,9 @@ public class AVPBlockTagProvider extends FabricTagProvider.BlockTagProvider {
             HumanPlasticBlocks.DYE_COLOR_TO_CUT_PLASTIC_SLAB,
             HumanPlasticBlocks.DYE_COLOR_TO_CUT_PLASTIC_STAIRS,
             HumanPlasticBlocks.DYE_COLOR_TO_FRAMED_PLASTIC,
+            HumanPlasticBlocks.DYE_COLOR_TO_PITTED_PLASTIC,
+            HumanPlasticBlocks.DYE_COLOR_TO_PITTED_PLASTIC_SLAB,
+            HumanPlasticBlocks.DYE_COLOR_TO_PITTED_PLASTIC_STAIRS,
             HumanPlasticBlocks.DYE_COLOR_TO_PLASTIC,
             HumanPlasticBlocks.DYE_COLOR_TO_PLASTIC_SLAB,
             HumanPlasticBlocks.DYE_COLOR_TO_PLASTIC_STAIRS